airflow/requirements.txt
pandas
numpy
psycopg2-binary
duckdb